+++ pkg/domain/service/place/service.go
@@ -41,14 +41,15 @@ func (s *service) Create(ctx context.Context, masterTx repository.MasterTx, name
 		UpdatedAt: &createdAt,
 	}
 	result, err := s.placeRepository.Insert(ctx, masterTx, place)
-	place.ID = result
-
 	if err != nil {
 		return nil, werrors.Stack(err)
 	}
+	place.ID = result
 	return model.ConvertToPlaceEntity(place), nil
 }
 
+// NOTE: 空値があった時、元データが消滅する。
+// NOTE: リクエストは、全フィールド埋める or 差分だけ
 func (s *service) Update(ctx context.Context, masterTx repository.MasterTx, placeID int, name string) (*entity.Place, error) {
 	place, err := s.placeRepository.SelectByID(ctx, masterTx, placeID)
 	if err != nil {
